Refactor libp2p configs
